services/doc-parser/requirements.txt
fastapi~=0.111
python-docx~=1.0
uvicorn[standard]~=0.30
